requirements.txt
# Core DL stack (CPU)
torch==2.3.1
transformers==4.41.2
sentence-transformers==2.7.0

# Data + utilities
datasets==2.20.0
jsonlines==4.0.0
tqdm==4.66.4
pandas==2.2.2
matplotlib==3.8.4
joblib==1.3.2
scikit-learn==1.4.2

# Retrieval extras
rank-bm25==0.2.2

# App
streamlit==1.36.0

faiss-cpu==1.8.0.post1

google-genai

